pyproject.toml
[project]
name = "dronecomm"
version = "0.1.0"
description = "Geometry-aware drone deployment simulation for disaster emergency communications"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "cma>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["dronecomm*"]
